+++ docs/zh/CHANGELOG.md
@@ -15,6 +15,7 @@
 #### 添加
 - 增加头文件 /usr/include/ulpatch/version.h。
 - rpmbuild.sh 添加 --nocheck 参数。
+- task: 添加C宏 `task_vdso_vma(task)`。
 #### Changed
 - 重命名 src/ulconfig 为 src/ulpconfig.sh.in。
 #### 弃用